+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Demo for grid `Sparklines` renderer in Portfolio example
 * Added SlackAlertService to post StatusMonitor and ClientError alerts to the XH slack.
+* Added two simple StatusMonitors, `metric1337Monitor` and `divideByZeroMonitor`, for testing purposes.
 
 ## v2.19.0 - 2022-07-30